requirements.txt
flask == 1.1.2
flask_ngrok == 0.0.25
nltk == 3.5
numpy == 1.19.5
pandas == 1.2.1
tensorflow-cpu == 2.4.0
